refactor:adapter rag storage and add integration documents.
